feat(client): pocket페이지 모달 통합
