+++ client/src/pocket/components/Modal/Template.tsx
@@ -1,22 +1,20 @@
 import * as Sandpack from '@codesandbox/sandpack-react';
 import { useActiveCode } from '@codesandbox/sandpack-react';
 import { Modal } from '@shared/components';
-import type { ModalInterface } from '@shared/contexts/ModalContext';
 import { localStorage } from '@shared/utils/localStorage';
 import { rem } from 'polished';
 import React, { useEffect, useState } from 'react';
 
-import useCode from '../../hooks/useCode';
 import usePushCode from '../../hooks/usePushCode';
 import * as style from './style.css';
 
-interface ModalContentProps {
+interface ModalContentTemplateProps {
   codeName?: string;
   useAnonymousMode?: boolean;
   closeModal?: () => void;
 }
 
-const ModalContent = (props: ModalContentProps) => {
+const ModalContentTemplate = (props: ModalContentTemplateProps) => {
   const { code } = useActiveCode();
   const [codeName, setCodeName] = useState('');
   const [useAnonymousMode, setUseAnonymousMode] = useState(false);
@@ -47,9 +45,9 @@ const ModalContent = (props: ModalContentProps) => {
     setCodeName(event.target.value);
 
   return (
-    <div className={style.editModalContent}>
-      <h1 className={style.editModalHeaderText}>새로운 코드 등록하기</h1>
-      <div className={style.editModalHeaderWrapper}>
+    <div className={style.modalContent}>
+      <h1 className={style.modalHeaderText}>새로운 코드 등록하기</h1>
+      <div className={style.modalHeaderWrapper}>
         <div style={{ flexGrow: 8 }}>
           <label className={style.label} htmlFor="codename">
             코드이름
@@ -59,7 +57,7 @@ const ModalContent = (props: ModalContentProps) => {
             id="codename"
             value={codeName}
             placeholder="코드이름을 지어주세요(ex. useCode)"
-            className={style.editModalTitleInput}
+            className={style.modalTitleInput}
             onChange={onChangeCodeName}
           />
         </div>
@@ -85,7 +83,7 @@ const ModalContent = (props: ModalContentProps) => {
       <Sandpack.SandpackLayout style={{ height: '100%' }}>
         <Sandpack.SandpackCodeEditor id="code" style={{ height: '100%', fontSize: rem(18) }} />
       </Sandpack.SandpackLayout>
-      <div className={style.editModalButtonContainer}>
+      <div className={style.modalButtonContainer}>
         <Modal.ConfirmButton text="수정하기" onConfirm={onConfirm} />
         <Modal.CancelButton text="닫기" onCancel={props.closeModal} />
       </div>
@@ -93,24 +91,4 @@ const ModalContent = (props: ModalContentProps) => {
   );
 };
 
-const CreateModal = ({ closeModal, targetId }: ModalInterface) => {
-  const { data: codeInfo } = useCode({ codeId: targetId });
-
-  return (
-    <div className={style.editModalContainer}>
-      <Sandpack.SandpackProvider
-        template="react-ts"
-        files={{ '/App.tsx': codeInfo?.code || '' }}
-        style={{ height: '100%' }}
-      >
-        <ModalContent
-          closeModal={closeModal}
-          codeName={codeInfo?.codeName}
-          useAnonymousMode={codeInfo?.isAnonymous}
-        />
-      </Sandpack.SandpackProvider>
-    </div>
-  );
-};
-
-export default CreateModal;
+export default ModalContentTemplate;
